C: change guesses to long

+++ C/src/bruteforce.c
@@ -4,11 +4,11 @@
 #include <stdlib.h>
 #include <time.h>
 
-int bruteforce_randomly(char password[], int min_length, int max_length) {
+long bruteforce_randomly(char password[], int min_length, int max_length) {
     srand(time(NULL));
     const char ALPHANUMERIC[62] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz1234567890";
     bool is_not_cracked = true;
-    int guesses = 1;
+    unsigned long guesses = 1;
 
     while(is_not_cracked) {
         int length = rand()% max_length;
